+++ spec/rubocop/cops/style/collection_methods_spec.rb
@@ -6,13 +6,22 @@ module Rubocop
   module Cop
     module Style
       describe CollectionMethods do
+        CollectionMethods.config = {
+          'PreferredMethods' => {
+            'collect' => 'map',
+            'inject' => 'reduce',
+            'detect' => 'find',
+            'find_all' => 'select'
+          }
+        }
+
         let(:cop) { CollectionMethods.new }
 
-        CollectionMethods::PREFERRED_METHODS.keys.each do |method|
+        CollectionMethods.preferred_methods.keys.each do |method|
           it "registers an offence for #{method} with block" do
             inspect_source(cop, ["[1, 2, 3].#{method} { |e| e + 1 }"])
             expect(cop.offences.size).to eq(1)
-            preferred_method = CollectionMethods::PREFERRED_METHODS[method]
+            preferred_method = CollectionMethods.preferred_methods[method]
             expect(cop.messages)
               .to eq(["Prefer #{preferred_method} over #{method}."])
           end
@@ -20,7 +29,7 @@ module Rubocop
           it "registers an offence for #{method} with proc param" do
             inspect_source(cop, ["[1, 2, 3].#{method}(&:test)"])
             expect(cop.offences.size).to eq(1)
-            preferred_method = CollectionMethods::PREFERRED_METHODS[method]
+            preferred_method = CollectionMethods.preferred_methods[method]
             expect(cop.messages)
               .to eq(["Prefer #{preferred_method} over #{method}."])
           end